Bug 1205930 - Tighten up warnings handling in media/omx-plugin/. r=gerald.
MetaData.h uses multi-char literals. This is kind of gross but probably safe
and hard to change, so this patch allows it while disallowing all other
compiler warnings. (This approach is already used in dom/media/moz.build.)

(An aside: I originally tried using |#pragma GCC ignoring -Wmultichar| to
temporarily disable the warnings in MetaData.h. But that works in gcc but not
in g++ because of https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53431!)

--HG--