Bug 1247346. Slot indices on IDL interface members that get a slot should be per-interface, not a single per-member number. r=peterv
This matters when a [Cached] attribute appears on an interface, call it C,
which is a consequential interface for both A and B.  There is no a priori
reason that A and B would have identical numbers of their own [Cached]
attributes, which will come before the members of consequential interfaces in
the member array, so in general the attribute may have different slot indices in A
and B.

The approach taken here is to keep things simple and make the slot index of an
interface member either None or a dict mapping interface name to the slot
index.  In the common case of only one interface being involved this is a
little wasteful, but it keeps things simple for consumers, and one extra dict
per [Cached] attribute is not a big deal in the grand scheme of the data
structures the Web IDL parser produces.  Another option would be to store None,
a number (if there is only one interface involved), or a dict (if there are
several interfaces involved).  If we did that, we'd probably want to add some
method to get the slot index for a given interface, because the logic for doing
that would get a bit more complicated.