Bug 1189671 - Fix getregistrations test. r=jgraham
Moves the test to .https so it actually works.

Switches to using promise_test like the current blink test since sequential promise is not implemented and not needed.

compares registrations by scope since getRegistrations() produces new objects every time.

Change the with_iframe() call for the remote frame registration to actually
wait till the remote register() finishes so there is a valid registration to
unregister() when the frame receives a message.