Bug 1209574 - Generate C++ name for any Java class in chain of declared classes; r=snorp
Right now, when we generate bindings for Java class A, and we encounter
a Java type B, we generate a corresponding C++ name only if A == B,
otherwise we generate a generic "jni::Object" C++ name. For example,

class Foo {
  class Bar {
    Foo getFoo(Bar bar);
  }
}

In C++, Foo.Bar.getFoo would become,

class Foo {
  class Bar {
    jni::Object::LocalRef getFoo(Bar::Param bar);
  };
};

This patch extends the code generator so that any Java class in the
chain of declared classes gets a corresponding C++ name. The above
example now becomes,

class Foo {
  class Bar {
    Foo::LocalRef getFoo(Bar::Param bar);
  };
};