Bug 1059312 - Fix highlighter offset after switching iframe context; =pbro
The auto-refresh highlighter base class was getting a reference to the window
object to use for highlighting at instanciation time.
This reference should be updated if a navigation of the highlight environment
occurs.

This commit maps the "win" property to a getter, so that win is always in sync
with the current highlight environment.