Bug 1259354 - Avoid "checking yasm version" being written out when yasm was not found. r=ted
The reason the "checking" string always appears is that @Depends
functions are always called, regardless of the value of the dependency.

This introduces a new decorator @depends_true, which works like
@Depends, but the decorated function is not called unless one of the
dependency value resolves to True.

The new decorator can also be used to replace many cases where we do
@Depends(foo)
def bar(foo):
    if foo:
        ...