Bug 1235908 - measure total event wait and process time. r=smaug
Native events come with timestamps indicating when they were created or
generated by the user. Using those, we can get a full picture of how long it
takes between the user trying to do something and us responding to it.

This is currently only for ports that populate WidgetEvent's timeStamp
(presently gtk, Windows).