Bug 1260891 - Acquire and release the lock when destorying an `ExclusiveData<T>`'s protected value; r=terrence a=kwierso
Even if there is not a traditional race between uses of the protected data and
the `ExclusiveData<T>`'s destruction, the thread running the destructor has no
guarantee that its cache is up to date with changes already made by other
threads. The new acquire and release surrounding the destructor ensures that the
destructor doesn't run on stale cache lines.