Bug 1238815 - Limit baseline script size on ARM. r=jandem
ARM branch instructions have a limited range of 32 MB, and our ARM
macroassembler doesn't mitigate that except by crashing when a branch goes out
of range.

Limit the size of scripts that baseline will attempt to compile on ARM so that
we are much less likely to hit the hard crash.